requirements.txt
# Core dependencies for the agent and LLM
langchain-google-genai
google-ai-generativelanguage
google-generativeai
pydantic  # Compatible version for structured output

# Browser automation with Playwright
playwright # Latest stable as of Oct 2024, adjust if needed

# Streamlit for the UI
streamlit  # Latest stable as of Oct 2024

# Asyncio helper for Streamlit compatibility
nest_asyncio==1.6.0

# Environment variable management
python-dotenv

# browser_use (assuming it's a custom or local package, not on PyPI)
# If it's on PyPI, uncomment and specify the version:
# browser_use==x.x.x
# If it's local, install it manually or via git (e.g., git+URL)